Fix ESM imports for vite dev
